[AOD-97] Let components include each other directly
While the majority of the functionality has become unavailable, the
specs are still run on all the components (`components/*.spec.js`).
Since a lot of the components import other components, but the index has
been stripped due to bundle size, we now resolve all component
dependencies directly to mitigate any errors in the specs.