[AOD-98] Implement guest tokens
This commit implements guest token negotiation with the Argu backend.
The internal communication system from FE to BE have been refactored to
provide better encapsulation and abstraction. All server-side calls to
the backend should be done with the provided API/SPI class.

The error handling system has also been refactored to accommodate the
new API workings. This should provide enough streamlining for the
foreseeable future. The error system has uses the errors for both FE-BE
and client-FE communication. The FE-BE errors are marked as `internal`
to differentiate between the two. These errors are not able to leave the
system since their request metadata might include elevated privileges (
e.g. an SPI token).

This also introduces changes in the webpack server bundle to leverage
the new ES2017 async/await feature to defer the async responsibility to
the underlying engine.