[AOD-196] Rework collapsible internals to be more safe
The previous implementation used unsafe methods (e.g. calling setState
in render) which can lead to unexpected behaviour.

We have to initialize a state in the constructor of Collapsible to use
`getDerivedStateFromProps`, though we don't update our own state, we do
have stateful behaviour in the form of calling functions as side effect,
so it seems to be fine from a theoretical standpoint.